añadida comprobacion de administrador a borrar servicio y usuario

+++ trunk/controlador/deleteservice.php
@@ -1,11 +1,16 @@
 <?php
- 
-$id = $_GET['id'];
- 
-$enlace = mysqli_connect('localhost', 'root', '', 'intime');
+ session_start();
 
-$query = "DELETE FROM servicio WHERE id_servicio = '$id'";
-mysqli_query($enlace, $query);
-header('location:../vista/serviciosadmin.php');
+require_once '../controlador/comprobar_login_usuario_admin.php';
+
+ if ((isset($_SESSION['login_admin'])) && ($_SESSION['login_admin'] == true)){
+	$id = $_GET['id'];
+	
+	$enlace = mysqli_connect('localhost', 'root', '', 'intime');
+	
+	$query = "DELETE FROM servicio WHERE id_servicio = '$id'";
+	mysqli_query($enlace, $query);
+	header('location:../vista/serviciosadmin.php');
+}
  ?>
